+++ Cargo.toml
@@ -18,6 +18,7 @@ rayon = "1.6.1"
 log = "0.4.17"
 crossbeam-epoch = "0.9.13"
 bincode = "1.3.3"
+anyhow = "1.0.69"
 
 [dev-dependencies]
 criterion = "0.4.0"
