--- experiments/tokio-async-shared-mutable-state/src/main.rs
@@ -1,6 +1,0 @@
-mod map;
-
-#[tokio::main]
-async fn main() {
-    println!("Hello, world!");
-}
