+++ src/concurrent/data_structures/maps/mod.rs
@@ -1,0 +1,1 @@
+mod simple_shared_map;
