+++ src/concurrent/data_structures/mod.rs
@@ -1,1 +1,2 @@
+pub mod maps;
 pub mod treiber_stack;
